refactor(gutter): use defaultlineheight to set gutter rect
